🚚 Move: `Bound` enum to `param::pagination` module [#63]
