+++ src/client/param/mod.rs
@@ -1,5 +1,6 @@
 //! Features for the parameters of the [`Client`](crate::client::Client) struct methods.
 
 pub mod news_stream;
+pub mod pagination;
 pub mod search_user;
 pub mod user_leaderboard;
